Add conjugation tables to pronouns
This also removes the now-redundant entries for conjugated pronoun
forms, because searching on them will just lead to the root form (for
example, searching for pxoengal should return oeng, with its conjugation
table).

Note that in the current version of Reykunyu, such conjugated pronouns
cannot be found yet, but this will be implemented in the future.